test: move changed to cli
