--- test/changed/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@vitest/test-changed",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "vitest": "workspace:*"
-  }
-}
