--- test/changed/vitest.config.js
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    include: ['tests/**/*.test.ts'],
-    testTimeout: 60_000,
-    hookTimeout: 60_000,
-  },
-})
